requirements.txt
arrow==1.3.0
asgiref==3.7.2
asttokens==2.3.0
backcall==0.2.0
binaryornot==0.4.4
black==23.11.0
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.3.1
click==8.1.7
cookiecutter==2.4.0
decorator==5.1.1
Django==4.2.4
django-browser-reload==1.12.0
django-extensions==3.2.3
django-filter==23.3
django-tailwind==3.6.0
djhtml==3.0.6
executing==1.2.0
flake8==6.1.0
idna==3.4
ipython==8.15.0
isort==5.12.0
jedi==0.19.0
Jinja2==3.1.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mypy-extensions==1.0.0
packaging==23.2
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==4.0.0
prompt-toolkit==3.0.39
psycopg2==2.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.11.1
pydocstyle==6.3.0
pyflakes==3.1.0
Pygments==2.16.1
python-dateutil==2.8.2
python-slugify==8.0.1
PyYAML==6.0.1
requests==2.31.0
rich==13.6.0
six==1.16.0
snowballstemmer==2.2.0
sqlparse==0.4.4
stack-data==0.6.2
text-unidecode==1.3
traitlets==5.9.0
types-python-dateutil==2.8.19.14
urllib3==2.0.7
wcwidth==0.2.6
